+++ jest.config.ts
@@ -15,10 +15,13 @@ const config: Config.InitialOptions = {
       "<rootDir>/packages/implementations/react/src/atomic/Box/src",
     "@rarui-react/button":
       "<rootDir>/packages/implementations/react/src/atomic/Button/src",
+
     "@rarui-react/link":
       "<rootDir>/packages/implementations/react/src/atomic/Link/src",
     "@rarui-react/text":
       "<rootDir>/packages/implementations/react/src/atomic/Text/src",
+    "@rarui-react/icon-button":
+      "<rootDir>/packages/implementations/react/src/atomic/IconButton/src",
     "@rarui-react/icon":
       "<rootDir>/packages/implementations/react/src/atomic/Icon/src",
     "@rarui-react/divider":