+++ packages/implementations/react/src/atomic/Toast/src/hooks/useToast/useToast.ts
@@ -1,0 +1,5 @@
+import { useContext } from "react";
+
+import { Context, ContextProps } from "../../context";
+
+export const useToast: () => ContextProps = () => useContext(Context);