+++ packages/implementations/react/CHANGELOG.md
@@ -2,11 +2,24 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
-## 2024-04-26 `1.12.0`
+## 2024-04-26 `1.14.0`
 
 #### 🎉 New features
 
 - Added new atomic `Divider` component. ([#28](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/28) by [@luiza.rodrigues](https://git.rarolabs.com.br/luiza.rodrigues))
+
+## 2024-04-18 `1.13.0`
+
+#### 🎉 New features
+
+- Added new atomic `Banner` component. ([#26](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/26) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+
+## 2024-04-25 `1.12.0`
+
+#### 🎉 New features
+
+- Added new atomic `Toast` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+- Added new atomic `IconButton` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@junior](https://git.rarolabs.com.br/junior))
 
 ## 2024-04-25 `1.11.0`